Fixes html tag mistmatch in dashbaord template
For some reason there seemed to be an errant closing </div> tag. It was not having an effect on display.

Also removed DataTables from #sync_history, as there was no element with id="sync_history", and when
i tried adding it, it errored with "DataTables is not defined" anyway.

The table of sync history looks fine right now, so scrapping it.